better split devtools containers
